… third_party/CMakeLists.txt
… third_party/cereal/archives/json.hpp
… third_party/cereal/cereal.hpp
… third_party/cereal/details/polymorphic_impl.hpp
… third_party/cereal/details/static_object.hpp
… third_party/cereal/external/LICENSE
… third_party/cereal/external/rapidjson/LICENSE
… third_party/cereal/external/rapidjson/document.h
… third_party/cereal/external/rapidjson/msinttypes/LICENSE
… third_party/cereal/external/rapidxml/license.txt
… third_party/cereal/external/rapidxml/manual.html
… third_party/cereal/types/memory.hpp