@GindaChen 2021-03-12
Master
```
------------------------------------------------------
Benchmark            Time             CPU   Iterations
------------------------------------------------------
ssb_query11   79109272 ns       931473 ns          100
ssb_query12    9767603 ns       605645 ns          963
ssb_query13    9391787 ns       659630 ns         1092
ssb_query21  127454356 ns      1615069 ns          100
ssb_query22  404438392 ns      7672487 ns           10
ssb_query23  102338986 ns      1083056 ns          100
ssb_query31  175337496 ns      1129197 ns          100
ssb_query32   94747767 ns      2073993 ns          100
ssb_query33   85415839 ns      1277084 ns          100
ssb_query34   17616428 ns       761411 ns          920
ssb_query41  161491972 ns      1155526 ns          100
ssb_query42  106896752 ns      1409256 ns          100
ssb_query43   80518691 ns      2006849 ns          100
```

This Branch
```
------------------------------------------------------
Benchmark            Time             CPU   Iterations
------------------------------------------------------
ssb_query11   78589504 ns       933356 ns          100
ssb_query12    9751740 ns       592478 ns         1009
ssb_query13    9430765 ns       636661 ns         1169
ssb_query21  128076171 ns      1576499 ns          100
ssb_query22  460332596 ns      8025089 ns           10
ssb_query23  102906605 ns      1057554 ns          100
ssb_query31  191480606 ns      1104036 ns          100
ssb_query32   95855053 ns      2073153 ns          100
ssb_query33   83495965 ns      1153445 ns          100
ssb_query34   22932665 ns       879605 ns          923
ssb_query41  166233661 ns      1093335 ns          100
ssb_query42  108904179 ns      1463760 ns          100
ssb_query43   80341318 ns      1947974 ns          100
```