@GindaChen 2021-03-13
Merge this as we have figured out all source of errors in Mac. Next step is to clean up our build script so that we are insulate from accidental dependencies that are not built upon the same compiler (see #99)